Add zk-index-make-buttons
Separate this into its own function and make interactive, in case someone
saves an index buffer and wants to re-make the buttons. They'll just need to
open the file, call this function, and enter zk-index-mode.

(Not advertising this possibility, since indexes aren't really meant to be
saved. Better to use zk-desktop for that.)