Refactor zk-file-p to improve performance
Based on profiling, file-truename is an expensive operation, and
file-in-directory-p calls it many times. Most of the time, callers of zk-file-p
are not all that interested in making sure the file is indeed in zk-directory,
just that it's a valid file that matches the zk-id-regexp. Functions that do
need to check can pass STRICT optional argument.