Optimize (eq 1 (length list)) throughout with (null (cdr list))
Replacing the (eq 1 (length return)) test with (cdr result) avoids
having to "walk" the length of the list each time. With a list of a list
of 2243 files, difference adds up to 0.43 seconds for 100,000
repetitions and to 5.55 seconds for 1,000,000 repetitions. Quite minor,
but since I already wrote the refactored version for zk--parse-file in
the comments here, it seemed like a good time to do it everywhere.