Split zk-index-refresh into command and internal function
This better encapsulates functionality, so that the internal function
zk-index--refresh that requires all the parameters doesn't need to fill in
missing parameters when called as a command. Both the function and the command
are now more modular, doing just one thing, so it will be easier to make changes
if need be without having to worry about affecting unrelated things (user
experience vs. internal function).