Fix odd behavior with zk--singleton-p
Donald Knuth's adage "premature optimization is the root of all evil" proves
right once again. The problem was the result of my trying to use the
newfangled (as of 25.1) `define-inline` rather than `defun` without
understanding what that entailed. Even as a normal function, `zk--singleton-p`
is $O(1)$ rather than $O(n)$ as with `(= 1 (length list))` (in practical terms,
0.03 seconds vs. 0.55 seconds for 1,000,000 repetitions over a list of 2423
files), so there was no point in trying to inline it to begin with.