Refactor zk--format and zk-format-function
I thought I was being clever in "overloading" `zk--format` to call
`zk-format-function` if it is set, and to do the default formatting otherwise.
However, when writing `zk-format-function` for my use-case, I realized it would
be useful to have an option of calling the default formatter in some cases.
While it is possible to do that by lexically clearing `zk-format-function`
before calling `zk--format`, that seems hacky:

``` emacs-lisp
(defun my-custom-zk-format (format id title)
  (if (not (some-test-p id))
      (let ((zk-format-function nil))
        (zk--format format id title))
    ;; do the formatting on my own
    ...))
```

With this refactoring, the above can become:

``` emacs-lisp
(defun my-custom-zk-format (format id title)
  (if (not (some-test-p id))
      (zk--format-function format id title)
    ;; do the formatting on my own
    ...))
```

There is no change in the behavior; the default formatter is just moved to a new
function, `zk--format-function`, while `zk--format` becomes a very simple macro
to call whatever `zk-format-function` is set to. In effect, the macro just acts
as syntactic sugar to avoid having to type the lengthy `(funcall
zk-format-function format id title)` all the time. Since the macro is so simple,
does not ever need to change, and the biggest downside of macros --- they can't
be used in `mapcar`, for example --- is not an issue given its three arguments,
there shoulnd't be any surprises.