Preserve custom sort by setting :sort to nil
The :sort parameter cannot a function, it can only be nil or non-nil;
so any use of 'consult--read' that wants to use a custom sort order
must sort candidates prior to calling 'consult--read', that is,
the sorting must be done in the function itself, not in completing-read,
via display-sort-function mechanism; so, setting :sort to nil
preserves that in-function pre-sorting. got it?