Add zk--formatter, to handle args from embark-select
'embark-select' passes lists of targets, so a few functions need to be able
to handle such lists, whether strings, singletons, or multi-item lists;
zk--formatter is a universal handler, which transformer any list to a list of
files, then produces a formatted list according to the format arg. This
formatted list can then be handled by another function, for example, the new
zk--formatted-string, which produces output appropriate for zk-insert-link
and zk-copy-link-and-title.

Ok. To see this in action, use embark-act-all to call zk-copy-link-and-title
on multiple targets, such as a narrowed group in the minibuffer.