Fix bug with zk-backlinks; add zk--posix-regexp to sanitize regexps
So far, functions like `zk--grep-file-list` have just passed Elisp-style regexp
to `grep` and hoped it works, which it does for zk-ID or other literal strings.
Changes to `zk-link-regexp` in PR #63 broke things because it introduced
explicit capture groups (`\(?1:...\)`), which is not supported by POSIX regexps
that `grep` uses.

The new function, `zk--posix-regexp`, does some basic conversion from
Elisp-style regexps to POSIX-style. There is package `pcre2el` that does it in a
more complete and sophisticated way, but for our purposes, this should be
enough.

Functions that pass regexps to `grep`, such as `zk--grep-file-list`,
`zk--grep-tag-list`, and `zk--grep-link-id-list` are updated to
sanitize ("POSIXize") the regexps they pass. The cost is minuscule ("Elapsed
time: 0.000379s" for 10,000 calls with a complex regexp).