Fix references to non-existing match group in zk-id-regexp
`zk-id-regexp` does not specify that the regexp should be wrapped in
`\\(...\\)`, but some code that matches against the regexp refers to
match-string 1 while other instances (e.g. `zk--id-at-point`) do not. Since
`zk-link-regexp` and `zk-file-name-regexp` add their own explicitly-numbered
capture groups, it's best to consistently refer to the entire match when
matching against `zk-id-regexp`.