Feat: Made zk-desktop-mode into a properly toggling minor mode
This allows the user an option to treat Zk-Desktop file as a regular text
file (in other words, without buttons, read-only text properties, or overlays).
Someone who uses org-mode could, for example, just set it up that the Zk-Desktop
files are annotated org-mode documents, with each item as a heading or a list
item by setting `zk-desktop-prefix` to `"** "` or `"- "`, respectively.