Speed up auto-scroll, dramatically
'zk-index-view-mode' should not run 'zk-make-link-buttons' because it
calls the fairly expensive 'zk--id-list'.

Fixed by let-binding 'zk-enable-link-buttons' to nil in 'zk-index-view-note'
function, and calling 'zk-make-link-buttons' when/if 'zk-index-view-mode' is
turned off.

The lagginess of auto-scroll has bugged me for a while, and I always thought
the culprit was loading org and org-indent. But, alas, it was my fault all
along... :/